OOPDEF-61: Updated ReviewPage component.
